adding default throttling values for both cosmosdb and sqlserver ; moved KnownDataStores so that it can be used by the throttling middleware as well as the web startup class
