+++ src/Microsoft.Health.Fhir.Api/Features/Throttling/ThrottlingMiddleware.cs
@@ -11,12 +11,15 @@ using System.Threading;
 using System.Threading.Tasks;
 using EnsureThat;
 using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.Health.Abstractions.Exceptions;
 using Microsoft.Health.Fhir.Api.Configs;
 using Microsoft.Health.Fhir.Api.Features.Headers;
 using Microsoft.Health.Fhir.Core.Configs;
+using Microsoft.Health.Fhir.Core.Features;
+using Microsoft.Health.Fhir.Core.Features.Operations;
 
 namespace Microsoft.Health.Fhir.Api.Features.Throttling
 {
@@ -39,6 +42,7 @@ namespace Microsoft.Health.Fhir.Api.Features.Throttling
         private const string ThrottledContentType = "application/json; charset=utf-8";
         private static readonly ReadOnlyMemory<byte> _throttledBody = CreateThrottledBody(Resources.TooManyConcurrentRequests);
 
+        private IConfiguration _baseConfiguration;
         private readonly RequestDelegate _next;
         private readonly ILogger<ThrottlingMiddleware> _logger;
         private readonly HashSet<(string method, string path)> _excludedEndpoints;
@@ -58,6 +62,7 @@ namespace Microsoft.Health.Fhir.Api.Features.Throttling
 
         public ThrottlingMiddleware(
             RequestDelegate next,
+            IConfiguration baseConfiguration,
             IOptions<ThrottlingConfiguration> throttlingConfiguration,
             IOptions<SecurityConfiguration> securityConfiguration,
             ILogger<ThrottlingMiddleware> logger)
@@ -67,6 +72,7 @@ namespace Microsoft.Health.Fhir.Api.Features.Throttling
             ThrottlingConfiguration configuration = EnsureArg.IsNotNull(throttlingConfiguration?.Value, nameof(throttlingConfiguration));
             EnsureArg.IsNotNull(securityConfiguration?.Value, nameof(securityConfiguration));
 
+            _baseConfiguration = EnsureArg.IsNotNull(baseConfiguration, nameof(baseConfiguration));
             _throttlingEnabled = throttlingConfiguration.Value.Enabled;
 
             _securityEnabled = securityConfiguration.Value.Enabled;
@@ -82,7 +88,19 @@ namespace Microsoft.Health.Fhir.Api.Features.Throttling
             }
 
             // snapshot the configuration values to reduce the number of instructions that need to execute in the lock.
-            _concurrentRequestLimit = configuration.ConcurrentRequestLimit;
+            if (_baseConfiguration["DataStore"].Equals(KnownDataStores.CosmosDb, StringComparison.OrdinalIgnoreCase))
+            {
+                _concurrentRequestLimit = configuration.ConcurrentRequestLimit == 0 ? (int)ThrottlingLimitDefault.Gen1 : Math.Max(configuration.ConcurrentRequestLimit, (int)ThrottlingLimitDefault.Gen1);
+            }
+            else if (_baseConfiguration["DataStore"].Equals(KnownDataStores.SqlServer, StringComparison.OrdinalIgnoreCase))
+            {
+                _concurrentRequestLimit = configuration.ConcurrentRequestLimit == 0 ? (int)ThrottlingLimitDefault.Gen2 : Math.Max(configuration.ConcurrentRequestLimit, (int)ThrottlingLimitDefault.Gen2);
+            }
+            else
+            {
+                _concurrentRequestLimit = configuration.ConcurrentRequestLimit;
+            }
+
             _maxMillisecondsInQueue = configuration.MaxMillisecondsInQueue;
             _maxQueueSize = _maxMillisecondsInQueue == 0 ? 0 : configuration.MaxQueueSize;
