+++ src/Microsoft.Health.Fhir.Core/Features/Operations/ThrottlingLimitDefault.cs
@@ -1,0 +1,13 @@
+﻿// -------------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+// -------------------------------------------------------------------------------------------------
+
+namespace Microsoft.Health.Fhir.Core.Features.Operations
+{
+    public enum ThrottlingLimitDefault
+    {
+        Gen1 = 15,
+        Gen2 = 75,
+    }
+}
